Rename to conflict with exported Member type
